requirements.txt
albumentations == 1.3.0
Flask == 2.0.1
torchvision == 0.10.0
geffnet == 1.0.2